[ANGLE] Allow envvar to gate ASSERT at runtime https://bugs.webkit.org/show_bug.cgi?id=281156 rdar://137613945
Reviewed by NOBODY (OOPS!).

Add a `AreAssertionsEnabled()` function to `log_utils.h` to check a
runtime envvar once, and cache said value. This will allow us to
modulate the ASSERTs crashing the process regardless of how the ANGLE
build was done.

The env var gating this is `ANGLE_DISABLE_ASSERTS`.

Additionally, the function is gated behind the `ANGLE_ENABLE_ASSERTS`
guard, so this will have no impact on production builds.

* Source/ThirdParty/ANGLE/src/common/log_utils.h:
(AreAssertionsEnabled):